requirements.txt
flask==2.0.1
requests==2.31.0
pillow==9.0.0
pytesseract==0.3.8
opencv-python==4.5.3.56
numpy==1.21.2
werkzeug==2.0.2
sqlalchemy==1.4.23
python-dotenv==0.19.0
pytest==6.2.5
gunicorn==21.2.0
Flask-WTF==1.1.1
WTForms==3.0.1
black==23.7.0
isort==5.12.0
gitpython==3.1.31